Move example scripts to Examples/Scripts
